refactor: split oversized cli parser files under structure limit
